Fixed warnings in CCTexturePVR

+++ libs/cocos2d/CCTexturePVR.m
@@ -303,7 +303,8 @@ typedef struct _PVRTexHeader
 			glTexImage2D(GL_TEXTURE_2D, i, internalFormat, width, height, 0, format, type, data);
 
 		if( i > 0 && (width != height || ccNextPOT(width) != width ) )
-			CCLOG(@"cocos2d: TexturePVR. WARNING. Mipmap level %u is not squared. Texture won't render correctly. width=%u != height=%u", i, width, height);
+			CCLOG( @"cocos2d: TexturePVR. WARNING. Mipmap level %u is not squared. Texture won't render correctly. width=%u != height=%u", 
+				  (unsigned int)i, (unsigned int)width, (unsigned int)height );
 		
 		err = glGetError();
 		if (err != GL_NO_ERROR)
